chips: rp2040: pio: retire the examples module
pio.rs ended in a private `mod examples` whose methods were all `pub`,
so it was API rather than examples. Its doc comment read "Used for the
examples in the pico explorer base main.rs file." Nothing in
boards/pico_explorer_base has referenced any of it for some time.

Eleven methods lived there. A grep across chips, boards, capsules,
kernel and libraries finds callers for three:

    pwm_program_init         boards via chips/rp2040/src/pio_pwm.rs
    spi_program_init         chips/rp2040/src/pio_spi.rs
    cyw43_spi_program_init   chips/rp2040/src/pio_gspi.rs

Move those three next to their callers, as free functions taking the
Pio by reference. cyw43_spi_program_init is the one that matters: it
configures drive strength, slew rate and Schmitt triggers for the
half-duplex gSPI bus, which the generic PIO API does not expose, and it
now sits with the driver that needs it rather than in the middle of a
peripheral driver.

Delete the other eight, which have no callers anywhere:
blinking_hello_program_init, blink_program_init, sideset_program_init,
hello_program_init, read_instr, read_sideset_reg, read_dbg_padout and
read_fdebug.

Delete `impl RPGpio { fn from_u32 }` with them. RPGpio is already
declared inside enum_from_primitive!, so FromPrimitive::from_u32 exists
and returns Option. The inherent method shadowed it with thirty numbered
arms and a wildcard that panicked, naming the offending value and the
valid range. Callers use expect rather than unwrap so the range survives
in the message; the value does not, since expect takes a &str.

pio.rs loses 272 lines and no longer imports RPGpio.

    board                 text     bss    stripped binary
    raspberry_pi_pico    98348   16528    byte-identical
    pico_explorer_base  102444   72668    byte-identical
    nano_rp2040_connect  94252   14368    byte-identical
    raspberry_pi_pico_w 352300   20724    differs, see below

Sections are unchanged on all four. Three link byte for byte. On
raspberry_pi_pico_w, the only board that uses the gSPI initialiser,
RPGpio::from_u32 disappears (44 bytes), the expect path brings in
Option::expect (16), expect_failed (24) and panic_display::<&str> (28),
and start grows 36 bytes where they inline. Alignment absorbs the
difference, so text is unchanged.

Nothing here was run on RP2040 hardware; the claims above are link and
size comparisons.
