+++ chips/rp2040/src/pio.rs
@@ -19,7 +19,7 @@ use kernel::utilities::registers::interfaces::{ReadWriteable, Readable, Writeabl
 use kernel::utilities::registers::{ReadOnly, ReadWrite, register_bitfields, register_structs};
 use kernel::{ErrorCode, debug};
 
-use crate::gpio::{GpioFunction, RPGpio, RPGpioPin};
+use crate::gpio::{GpioFunction, RPGpioPin};
 
 const NUMBER_STATE_MACHINES: usize = 4;
 const NUMBER_INSTR_MEMORY_LOCATIONS: usize = 32;
@@ -1671,277 +1671,5 @@ impl Pio {
             state_machine.config(&default_config);
         }
         self.clear_instr_registers()
-    }
-}
-
-mod examples {
-    use kernel::hil::gpio::Configure;
-
-    use super::{
-        DBG_PADOUT, FDEBUG, Pio, RPGpio, RPGpioPin, Readable, SMNumber, SMx_EXECCTRL, SMx_INSTR,
-        SMx_PINCTRL, StateMachineConfiguration, debug,
-    };
-
-    impl RPGpio {
-        fn from_u32(value: u32) -> RPGpio {
-            match value {
-                0 => RPGpio::GPIO0,
-                1 => RPGpio::GPIO1,
-                2 => RPGpio::GPIO2,
-                3 => RPGpio::GPIO3,
-                4 => RPGpio::GPIO4,
-                5 => RPGpio::GPIO5,
-                6 => RPGpio::GPIO6,
-                7 => RPGpio::GPIO7,
-                8 => RPGpio::GPIO8,
-                9 => RPGpio::GPIO9,
-                10 => RPGpio::GPIO10,
-                11 => RPGpio::GPIO11,
-                12 => RPGpio::GPIO12,
-                13 => RPGpio::GPIO13,
-                14 => RPGpio::GPIO14,
-                15 => RPGpio::GPIO15,
-                16 => RPGpio::GPIO16,
-                17 => RPGpio::GPIO17,
-                18 => RPGpio::GPIO18,
-                19 => RPGpio::GPIO19,
-                20 => RPGpio::GPIO20,
-                21 => RPGpio::GPIO21,
-                22 => RPGpio::GPIO22,
-                23 => RPGpio::GPIO23,
-                24 => RPGpio::GPIO24,
-                25 => RPGpio::GPIO25,
-                26 => RPGpio::GPIO26,
-                27 => RPGpio::GPIO27,
-                28 => RPGpio::GPIO28,
-                29 => RPGpio::GPIO29,
-                _ => panic!(
-                    "Unknown value for GPIO pin: {} (should be from 0 to 29)",
-                    value
-                ),
-            }
-        }
-    }
-
-    impl Pio {
-        // # Examples
-        /// Used for the examples in the pico explorer base main.rs file.
-        pub fn blinking_hello_program_init(
-            &self,
-            sm_number: SMNumber,
-            pin: u32,
-            config: &StateMachineConfiguration,
-        ) {
-            let sm = &self.sms[sm_number as usize];
-            sm.config(config);
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(pin)));
-            sm.set_enabled(false);
-            sm.set_pins_dirs(pin, 1, true);
-            sm.set_set_pins(pin, 1);
-            sm.init();
-            sm.set_enabled(true);
-        }
-
-        pub fn blink_program_init(
-            &self,
-            sm_number: SMNumber,
-            pin: u32,
-            config: &StateMachineConfiguration,
-        ) {
-            let sm = &self.sms[sm_number as usize];
-            sm.config(config);
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(pin)));
-            sm.set_enabled(false);
-            sm.set_pins_dirs(pin, 1, true);
-            sm.set_set_pins(pin, 1);
-            sm.init();
-            sm.set_enabled(true);
-        }
-
-        pub fn sideset_program_init(
-            &self,
-            sm_number: SMNumber,
-            pin: u32,
-            config: &StateMachineConfiguration,
-        ) {
-            let sm = &self.sms[sm_number as usize];
-            sm.config(config);
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(pin)));
-            self.gpio_init(&RPGpioPin::new(RPGpio::GPIO7));
-            sm.set_enabled(false);
-            sm.set_pins_dirs(pin, 1, true);
-            sm.set_pins_dirs(7, 1, true);
-            sm.set_set_pins(pin, 1);
-            sm.set_side_set_pins(7, 1, false, true);
-            sm.init();
-            sm.set_enabled(true);
-        }
-
-        pub fn hello_program_init(
-            &self,
-            sm_number: SMNumber,
-            pin1: u32,
-            pin2: u32,
-            config: &StateMachineConfiguration,
-        ) {
-            let sm = &self.sms[sm_number as usize];
-            // This is used to turn on specifically GPIOs 6 and 7 - LSB is for GPIO 0, next bit is for GPIO 1 etc.
-            let turn_on_gpio_6_7 = 0b11000000;
-            sm.config(config);
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(pin1)));
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(pin2)));
-            sm.set_enabled(false);
-            sm.set_pins_dirs(pin1, 1, true);
-            sm.set_pins_dirs(pin2, 1, true);
-            sm.init();
-            sm.set_enabled(true);
-            let _ = sm.push_blocking(turn_on_gpio_6_7);
-            let _ = sm.push_blocking(0);
-        }
-
-        pub fn pwm_program_init(
-            &self,
-            sm_number: SMNumber,
-            pin: u32,
-            pwm_period: u32,
-            config: &StateMachineConfiguration,
-        ) {
-            let sm = &self.sms[sm_number as usize];
-            // "pull" command created by pioasm
-            let pull_command = 0x8080_u16;
-            // "out isr, 32" command created by pioasm
-            let out_isr_32_command = 0x60c0_u16;
-            sm.config(config);
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(pin)));
-            sm.set_enabled(false);
-            sm.set_pins_dirs(pin, 1, true);
-            sm.set_side_set_pins(pin, 1, false, true);
-            sm.init();
-            let _ = sm.push_blocking(pwm_period);
-            sm.exec(pull_command);
-            sm.exec(out_isr_32_command);
-            sm.set_enabled(true);
-        }
-
-        pub fn spi_program_init(
-            &self,
-            sm_number: SMNumber,
-            clock_pin: u32,
-            in_pin: u32,
-            out_pin: u32,
-            config: &StateMachineConfiguration,
-        ) {
-            let sm = &self.sms[sm_number as usize];
-            sm.config(config);
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(clock_pin)));
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(in_pin)));
-            self.gpio_init(&RPGpioPin::new(RPGpio::from_u32(out_pin)));
-            sm.set_enabled(false);
-
-            // Important: make the sideset pin an output pin then make it a side set pin
-            // and make sure to do all of this BEFORE setting the outpin as an outpin
-            sm.set_out_pins(clock_pin, 1);
-            sm.set_pins_dirs(clock_pin, 1, true);
-            sm.set_side_set_pins(clock_pin, 1, false, false); // Do not switch pin dirs again, it will mess with the output settings
-
-            sm.set_pins_dirs(out_pin, 1, true);
-            sm.set_out_pins(out_pin, config.out_pins_count);
-
-            sm.init();
-            sm.clear_fifos();
-            sm.set_enabled(true);
-        }
-
-        pub fn cyw43_spi_program_init(
-            &self,
-            sm_number: SMNumber,
-            clock_pin: u32,
-            dio_pin: u32,
-            config: &StateMachineConfiguration,
-        ) {
-            let sm = &self.sms[sm_number as usize];
-            sm.set_enabled(false);
-            sm.config(config);
-            let clock_pin_handle = RPGpioPin::new(RPGpio::from_u32(clock_pin));
-            let dio_pin_handle = RPGpioPin::new(RPGpio::from_u32(dio_pin));
-            self.gpio_init(&clock_pin_handle);
-            self.gpio_init(&dio_pin_handle);
-
-            dio_pin_handle.set_floating_state(kernel::hil::gpio::FloatingState::PullNone);
-            dio_pin_handle.set_schmitt(true);
-            self.set_input_sync_bypass(&dio_pin_handle, true);
-            dio_pin_handle.set_drive_strength(crate::gpio::DriveStrength::Drive12ma);
-            dio_pin_handle.set_slew_rate(crate::gpio::SlewRate::Fast);
-            dio_pin_handle.activate_pads();
-
-            clock_pin_handle.set_drive_strength(crate::gpio::DriveStrength::Drive12ma);
-            clock_pin_handle.set_slew_rate(crate::gpio::SlewRate::Fast);
-            clock_pin_handle.set_floating_state(kernel::hil::gpio::FloatingState::PullNone);
-            clock_pin_handle.set_schmitt(true);
-            clock_pin_handle.set_slew_rate(crate::gpio::SlewRate::Slow);
-            clock_pin_handle.activate_pads();
-
-            sm.set_pins_dirs(dio_pin, 1, true);
-            sm.set_pins_dirs(clock_pin, 1, true);
-
-            sm.set_pins(&[&clock_pin_handle, &dio_pin_handle], false);
-
-            sm.init();
-            sm.clear_fifos();
-            sm.set_enabled(true);
-        }
-
-        // # Debugging
-        /// Returns current instruction running on the state machine.
-        pub fn read_instr(&self, sm_number: SMNumber) -> u32 {
-            self.registers.sm[sm_number as usize]
-                .instr
-                .read(SMx_INSTR::INSTR)
-        }
-
-        pub fn read_sideset_reg(&self, sm_number: SMNumber) {
-            debug!(
-                "{}",
-                self.registers.sm[sm_number as usize]
-                    .pinctrl
-                    .read(SMx_PINCTRL::SIDESET_COUNT)
-            );
-            debug!(
-                "{}",
-                self.registers.sm[sm_number as usize]
-                    .execctrl
-                    .read(SMx_EXECCTRL::SIDE_EN)
-            );
-            debug!(
-                "{}",
-                self.registers.sm[sm_number as usize]
-                    .execctrl
-                    .read(SMx_EXECCTRL::SIDE_PINDIR)
-            );
-            debug!(
-                "{}",
-                self.registers.sm[sm_number as usize]
-                    .pinctrl
-                    .read(SMx_PINCTRL::SIDESET_BASE)
-            );
-        }
-
-        pub fn read_dbg_padout(&self) -> u32 {
-            self.registers.dbg_padout.read(DBG_PADOUT::DBG_PADOUT)
-        }
-
-        pub fn read_fdebug(&self, tx: bool, stall: bool) -> u32 {
-            if tx {
-                if stall {
-                    self.registers.fdebug.read(FDEBUG::TXSTALL)
-                } else {
-                    self.registers.fdebug.read(FDEBUG::TXOVER)
-                }
-            } else if stall {
-                self.registers.fdebug.read(FDEBUG::RXSTALL)
-            } else {
-                self.registers.fdebug.read(FDEBUG::RXUNDER)
-            }
-        }
     }
 }
